Implementação inicial do gerenciador básico de permissões.
Melhorias na classe de conexão - adiciona insert e update genéricos.
Ajustes e melhorias na funcionalidade de atualizar os dados de cadastro no primeiro acesso à conta.

+++ Service/DeletePerson.php
@@ -2,10 +2,25 @@
 
 require_once '../config.php';
 require_once DIR . 'Repositorys/PersonRepository.php';
+require_once DIR . 'Repositorys/UserRepository.php';
 
 requireLogin();
 
+$module = new Module('excluir pessoas', 2);
+$result = Authenticator::verifyPermission($module);
+if ($result !== true) {
+	echo json_encode($result);
+	exit;
+}
+
 try {
+	echo json_encode(deletePersons());
+} catch(Exception $e) {
+	http_response_code(400);
+	echo $e->getMessage();
+}
+
+function deletePersons() {
 	$method = $_SERVER['REQUEST_METHOD'];
 	if ($method === 'DELETE') {
 		parse_str(file_get_contents('php://input'), $_DELETE);
@@ -13,18 +28,30 @@ try {
 
 		if (isset($ids) && !empty($ids)) {
 			$personRepository = new PersonRepository();
-			foreach ($ids as $id) {
-				$personRepository->deletePersonUser($id);
+			$userRepository = new UserRepository();
+			$deletingAdmins = false;
+
+			foreach ($ids as $i => $id) {
+				if ($userRepository->isAdministrator($id)) {
+					$deletingAdmins = true;
+					unset($ids[$i]);
+					continue;
+				} else {
+					$personRepository->deletePersonUser($id);
+				}
 			}
 
-			echo json_encode($personRepository->batchDelete($ids));
+			if (!empty($ids)) {
+				return $personRepository->batchDelete($ids);
+			}
+
+			if ($deletingAdmins) {
+				return new ErrorObj(400, 'Não é possível excluir a conta de outros administradores. Somente o dono da conta pode realizar essa ação.');
+			}
 		} else {
-			echo json_encode(new ErrorObj(400, 'Nenhuma pessoa foi selecionada.'));
+			return new ErrorObj(400, 'Nenhuma pessoa foi selecionada.');
 		}
 	} else {
-		echo json_encode(new ErrorObj(400, 'DELETE requested from a wrong method: "' . $method . '"'));
+		return new ErrorObj(400, 'DELETE requested from a wrong method: "' . $method . '"');
 	}
-} catch(Exception $e) {
-	http_response_code(400);
-	echo $e->getMessage();
 }
